requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0
httpx>=0.25.0
psycopg2-binary>=2.9.9
pdfplumber>=0.10.0
Pillow>=10.0.0
python-multipart>=0.0.6
